Always log an error when static graph-based restore fails
https://github.com/NuGet/Home/issues/12372
